pyproject.toml
[project]
name = "claude-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "claude-agent-sdk>=0.1.6",
    "python-dotenv>=1.2.1",
]
